+++ src_ts/src/kimi_k2_6/client.ts
@@ -40,6 +40,7 @@ import { fixOpenrouterUsageMetadata } from "../utils";
  */
 export class KimiK2_6Client extends LLMClient {
   protected _model: string;
+  protected _requiresNonThinkingOutput = true;
   private _client: OpenAI;
 
   /**