+++ src_ts/src/openai/client.ts
@@ -39,6 +39,7 @@ import { fixOpenrouterUsageMetadata } from "../utils";
  */
 export class OpenaiClient extends LLMClient {
   protected _model: string;
+  protected _requiresNonThinkingOutput = true;
   private _client: OpenAI;
 
   /**